Cleaner refactoring to determineManifestConversion
